requirements.txt
streamlit
pandas
python-dotenv
dashscope
langchain
langchain-community
requests